[VOID] TrackMap: Introduced a TrackMap which saves track items added to a track
The sole purpose of the track map is to reduce the time complexity for querying a track item at any given track frame.

Since the track items may or maynot have similar original ranges the track map allows query based on the timeline range of the track bifurcating the query into smalled based on start frame of the track item and then considering the range of the underlying media.

Viewer buffer also gets a locally cached track item which further reduces the queries to the underlying track or sequence making the whole play operation much optimised in terms of time complexity.